nvme: allow a debug module parameter to enable LBAs > PAGE_SIZE
As discused in preperation for LSFMM [0] enabling LBAs > PAGE_SIZE
triggers a crash with a null pointer dereference. If we want to get to
the poin to enable an NVMe device with a larger logical block size
greater than PAGE_SIZE we need to work out the kinks to support this.

Enable a module parameter for nvme_core to enable experimentation with
these types of devices.

[0] https://lkml.kernel.org/r/ZAJvu2hZrHu816gj@kbusch-mbp.dhcp.thefacebook.com

[pankaj: Added an extra conditional to not allow if BUFFER_HEAD is
enabled]

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>
Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>