xfs: fix min order usage and use MAX_PAGECACHE_ORDER
Enabling a block size > PAGE_SIZE is only possible if we can
ensure that the filesystem allocations for the block size is treated
atomically and we do this with the min order folio requirement for the
inode. This allows the page cache to treat this inode atomically even
if on the block layer we may treat it separately.

For instance, on x86 this enables eventual usage of block size > 4k
so long as you use a sector size set of 4k.

Fixes: ("xfs: set minimum order folio for page cache based on blocksize")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>