filemap: fix index used on filemap_get_pages() if not in page cache
We create a folio with filemap_create_folio() if the first batch
read with filemap_get_read_batch() fails on the page cache. When
we do this we use an unaligned index with iocb->ki_pos >> PAGE_SHIFT.
Fix this to use the aligned index. We had obtained that already
in an older commit, "filemap: align filemap_get_pages to align with
min order folio"

TODO: fold this in.
NOTE: this makes no progress on our fsx ops either.

Fixes: "filemap: align filemap_get_pages to align with min order folio"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>