mm/readahead.c: use min order folio multiplier on force_page_cache_ra()
We use a forced readahead with force_page_cache_ra() when readahead
is disabled and so the only the requested range is read. On 4k min
order systems that would issue 2 MiB read at a time up to the requested
read. Use a similar multiplier based on the min order so to enable
a bigger reads. So on a 16k min order system that's 8 MiB at a time.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>