page-cache: fix early rounding down on __filemap_get_folio()
iomap_write_end() uses pagecache_isize_extended() call when the size
increase. This in turn uses find_lock_page(), and the index needs to be
rounded down to the min order.

This helper is also used by truncate_setsize() and that is used even
by XFS directly. They call call pagecache_get_page() but this in turn
calls __filemap_get_folio() so the fix is actually to have rounded
down earlier on __filemap_get_folio() !

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>