mm/filemap.c: round down on filemap_get_folios_tag()
write back write_cache_pages() calls filemap_get_folios_tag() without
the indexes rounded down. Fix this. The start index is passed
as a pointer and we eventually update it with with the folio->index.

There's some fixation of math at the end of filemap_get_folios_tag()
but that seems fine.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>