zram: add alloc_block_bdev_range() and free_block_bdev_range()
Add [alloc|free]_block_bdev_range() which accepts number of blocks to
allocate or free from the block bitmap.

TODO: how alloc we reduce the number of IO when not able to allocate a
big chunk
TODO: how atomic is not needed for alloc function

This is useful when we want to send larger IOs to the backing dev.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>