mm/readahead.c: round down readahead sizes
Previously on commit titled ("mm/readahead.c: adjust readahead window to
account for min order") we had already accounted the min order on the
readahead window size. But that's not enough because sometimes we bump
the readahead window depending on the prior index and also sometimes
depending on the size of the request passed.

Fix all this and sprinkle some VM_BUG_ON() to make our goal clear.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>