page-cache: folio order allocation is at least min_order
The patch titled ("page-cache: ensure allocations are aligned to high
order") added an alignment check loop on __filemap_get_folio() and
to ensure we align the order used for a folio allocation. That is,
the check verifies the index we are on will align with a high order
folio. The check however ignores index 0 special case where anything
checked against it will be true and so the order could be 0 due to
earlier subtractions on the order. Fix this.

This should be folded in. A more clever loop should be used.

This gets us up to 13926 fsx ops with both:

./ltp/fsx    -d /media/test/foo
./ltp/fsx -W -d /media/test/foo

Fixes: ("page-cache: ensure allocations are aligned to high order"
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>