mm/truncate.c: replace min order alignment check
The current check was looking at the current index order which could
be different if we last had a large folio on the index. Use the min
order instead.

Yes, large folios should be aligned to their respective index, but we
can go bananas with those checks with a bit in more places once we
are certain the folio order didn't get reduced from the original
first allocation.

ie, if we truncate a large folio we'll likely want to always ensure
the truncation for that folio will be aligned even if it is larger
than the min order.

This takes us up to the million on fsx ops.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>