splice: don't check for uptodate if partially uptodate is impl
HACK! Remove before upstream.

When we do splice write, we copy from read buffer and it checks for the
uptodate flag in `pipe_buf_confirm` and that is not set for partial
folio and we return `EIO` from `page_cache_pipe_buf_confirm`. This issue
does not happen when we use `-X` because we come through the fault path
(because of readahead) and set the whole huge page with uptodate flag set.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>