readahead: allocate folios with mapping_min_order in ra_(unbounded|order)
page_cache_ra_unbounded() was allocating single pages (0 order folios)
if there was no folio found in an index. Allocate mapping_min_order folios
as we need to guarantee the minimum order if it is set.

page_cache_ra_unbounded() assumes that ractl->_index has already been
aligned to mapping_min_order. So when read_pages() is triggered if a page
is already present, check for truncation and move the ractl->_index by
mapping_min_nrpages if that folio was truncated. This is done to ensure
we keep the alignment requirement while adding a folio to the page cache.

page_cache_ra_order() tries to allocate folio to the page cache with a
higher order if the index aligns with that order. Modify it so that the
order does not go below the min_order requirement of the page cache.
This function also assumes that the ractl->_index is aligned to
min_order, and by allocating at least min_order folios, we guarantee the
alignment of folios added in this function.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>