readahead: align ractl index to min_order before adding pages to cache
page_cache_ra_order() and do_page_cache_ra() will be adding pages to
the page cache before reading it. As page cache requires the page index
to be aligned to its order, and to guarantee a folio to be mapping
min_order number of pages, align the ractl->_index to min_order.

Signed-off-by: Pankaj Raghav <p.raghav@samsung.com>